TUC 2019 Lightning Round Package

+++ mathtools_installer.ps1
@@ -1,6 +1,6 @@
 $PSVersionTable.PSVersion
 $ErrorActionPreference = "Continue"
-$global:logfile = "c:\ops\logs\mathtools_upgrade_output.txt"
+$global:logfile = "$PSScriptRoot\logs\mathtools_install_output.txt"
 Start-Transcript -path $logfile -Append
 Write-Host "Alpha"
 #check if InDesign CC version exists, and figure out the installation directory
@@ -15,11 +15,10 @@ $ids2018 = "Registry::HKEY_CLASSES_ROOT\CLSID\{74812DB7-FA97-43E0-97F5-87D1E47B7
 $ids2017 = "Registry::HKEY_CLASSES_ROOT\CLSID\{C62D9F67-2815-4C5D-9754-5CEAA121CDD8}\LocalServer32"
 #
 ## Edit these as needed ###
-$global:mathtools_version = "3_0_1_058" #Example: "3_0_1_055"
-$global:mathtools_url = "http://movemen.com/files/downloads/mtv3/058" #do not add trailing slash, example: "http://movemen.com/files/downloads/mtv3/055"
+$global:mathtools_version = "3_0_1_058" #Example: "3_0_1_058"
+$global:mathtools_url = "http://movemen.com/files/downloads/mtv3/058" #do not add trailing slash, example: "http://movemen.com/files/downloads/mtv3/058"
 $idsVersions = @("$ids2019","$ids2018","$ids2017") #example, "$ids2019","$ids2018","$ids2017"
 #
-
 
 ######## FUNCTION ####################
 function downloadMathTools($YYYY)
@@ -27,7 +26,7 @@ function downloadMathTools($YYYY)
   Write-Host "DEBUG: Downloading MathTools for InDesign CC" $YYYY
   $mathtools = "MathToolsEESrv-$mathtools_version-CC-$YYYY-WIN64.zip"
   $url = "$mathtools_url/$mathtools"
-  $global:output_dir = "c:\ops\temp\mathtools\$YYYY\"
+  $global:output_dir = "$PSScriptRoot\staging\mathtools\$YYYY\"
   $output = "$output_dir\$mathtools"
   $start_time = Get-Date
 
@@ -38,28 +37,15 @@ function downloadMathTools($YYYY)
 }
 
 ######## FUNCTION ####################
-function backupMathToolsLicense
-{
-  Write-Host "DEBUG: Copying license file from:" $ids_path_dir
-  New-Item -ItemType directory -Path $output_dir\lic -Force
-  Copy-Item $ids_path_dir\Plug-Ins\movemen\lic\*.lic -Destination $output_dir\lic -Force
-
-}
-
-######## FUNCTION ####################
-function updateMathTools($YYYY)
+function installMathTools($YYYY)
 {
   #call function
     Write-Host "DEBUG: Stopping InDesign CC $YYYY"
     Stop-Service -Name "InDesignServerService x64"
 
   #call function
-    Write-Host "DEBUG: Backing up MathTools License $YYYY"
-    backupMathToolsLicense
 
     #main body
-    #remove old MathTools
-    Remove-Item $ids_path_dir\Plug-Ins\movemen -Force -Recurse
 
     #prepare new MathTools
     #debug
@@ -69,11 +55,26 @@ function updateMathTools($YYYY)
 
     #production
     Microsoft.PowerShell.Archive\Expand-Archive -Path $output_dir\MathToolsEESrv-$mathtools_version-CC-$YYYY-WIN64.zip -DestinationPath $ids_path_dir\Plug-Ins -Force
-    Copy-Item  $output_dir\lic\*.lic -Destination $ids_path_dir\Plug-Ins\movemen\lic -Force
+    #if a license exists (maybe this is a reinstall?) then copy it in.
+    If (Test-Path $output_dir\lic\mt.MathToolsV2.lic)
+        {
+
+            Write-Host "Previous license found. Restoring."
+            Copy-Item  $output_dir\lic\*.lic -Destination $ids_path_dir\Plug-Ins\movemen\lic -Force
+        }
+
+    Else
+        {
+
+            Write-Host "Did not find existing MathTools Licenses."
+
+         }
+
 
   #call function
     Write-Host "DEBUG: Starting InDesign CC $YYYY"
     Start-Service -Name "InDesignServerService x64"
+
 
 }
 
@@ -85,15 +86,15 @@ function detectMathTools($YYYY)
 
 If (Test-Path $ids_path_dir\Plug-Ins\movemen)
     {
-        Write-Host "MathTools was found"
-        updateMathTools "$YYYY"
 
+        Write-Host "MathTools was found.  Skipping installation."
     }
 
 Else
     {
 
-        Write-Host "Did not find MathTools on this attempt."
+        Write-Host "Did not find MathTools on this attempt. Proceeding with installation."
+        installMathTools "$YYYY"
 
      }
 
@@ -113,17 +114,17 @@ Else
 
  foreach ($idsVer in $idsVersions) {
    "$idsVer = " + $idsVer.length
-   Write-Host "Trying InDesign:" $idsVer
+    Write-Host "Trying InDesign:" $idsVer
 
     If (Test-Path $idsVer)
         {
             Write-Host "$idsFound"
             $ids_path_exe = (Get-ItemProperty -LiteralPath "$idsVer").'(default)'
-            Write-Host "DEBUG: IDS EXE" $ids_path_exe
+            Write-Host "DEBUG:" $ids_path_exe
             $global:ids_path_dir = Split-Path -Path $ids_path_exe
-            Write-Host "DEBUG: IDS DIRECTORY" $ids_path_dir
+            Write-Host "DEBUG:" $ids_path_dir
             $global:idsYYYY = $ids_path_dir.substring($ids_path_dir.length - 4)
-            Write-Host "DEBUG: IDS YEAR" $idsYYYY
+            Write-Host "DEBUG:" $idsYYYY
 
             downloadMathTools "$idsYYYY"
